+++ tests/testthat/test-sort.R
@@ -26,12 +26,12 @@ test_that("sort1: Simple proc_sort test works.", {
 })
 
 
-test_that("sort2: Simple proc_sort with var works.", {
+test_that("sort2: Simple proc_sort with keep works.", {
 
 
 
   res <- proc_sort(prt, by = c("sex", "count"),
-                   var = c("sex", "count", "internship"))
+                   keep = c("sex", "count", "internship"))
 
   res
 
@@ -59,7 +59,7 @@ test_that("sort3: proc_sort defaults work.", {
 test_that("sort4: nodupkey work.", {
 
   res <- proc_sort(prt, by = c("sex", "internship"),
-                   var = c("sex", "internship"),
+                   keep = c("sex", "internship"),
                    nodupkey = TRUE)
 
   res
@@ -68,5 +68,76 @@ test_that("sort4: nodupkey work.", {
   expect_equal(ncol(res), 2)
 
 })
+
+
+test_that("sort5: proc_sort works with single sort variable.", {
+
+
+
+  res <- proc_sort(prt, by = c( "count"))
+
+  res
+
+  expect_equal(res[1, "count"], 10)
+  expect_equal(res[8, "count"], 53)
+
+
+})
+
+
+test_that("sort6: proc_sort works with unquoted variables.", {
+
+
+
+  res <- proc_sort(prt, by = sex, keep = v(sex, count))
+
+  res
+
+  expect_equal(ncol(res), 2)
+  expect_equal(res[4, 1], "boys")
+  expect_equal(res[8, 1], "girls")
+
+
+})
+
+
+test_that("sort7: single keep returns data frame.", {
+
+
+
+  res <- proc_sort(prt, by = count, keep = count)
+
+  res
+
+  expect_equal("data.frame" %in% class(res), TRUE)
+  expect_equal(ncol(res), 1)
+
+
+})
+
+
+test_that("sort8: proc_sort returns tibble.", {
+
+  prt2 <- as_tibble(prt)
+
+  res <- proc_sort(prt2, by = sex, keep = v(sex, count))
+
+  res
+
+  expect_equal(ncol(res), 2)
+  expect_equal("tbl_df" %in% class(res), TRUE)
+
+
+})
+
+test_that("sort8: order_df works as expected.", {
+
+
+expect_equal(1, 1)
+
+
+})
+
+